% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killconf.R
\name{killconf}
\alias{killconf}
\title{Function to perform sensitivity analysis of 'killer' confounders
This function takes a data frame, and d and s values as arguments, as defined by Ichino to calculate ATT, SE and ranges for lambda and gamma}
\usage{
killconf(
  data,
  treatment,
  indvar,
  outvar,
  categorical = TRUE,
  d,
  s,
  pu = 0.4,
  R = 1000,
  workers = 6,
  digits = 3,
  distance = "logit",
  caliperval = 0.1
)
}
\arguments{
\item{data}{a dataframe}

\item{treatment}{(\code{character}) Name of treatment variable}

\item{indvar}{(\code{character}) Vector of names of covariate variables}

\item{outvar}{(\code{character}) Name of the outcome variable}

\item{categorical}{(\code{boolean}) TRUE if categorical, FALSE if not}

\item{d}{(\code{numeric}) Value of the outcome effect of U in the absence of treatment}

\item{s}{(\code{numeric}) Value of the effect of U on the treatment selection}

\item{pu}{(\code{numeric}) Probability of U=1}

\item{R}{(\code{numeric}) Number of simulations}

\item{workers}{(\code{numeric}) Number of workers for parallel processing}

\item{digits}{(\code{numeric}) Number of digits}

\item{distance}{(\code{character}) Type of the distance; "logit" by default}

\item{caliperval}{(\code{numeric}) Value of the caliper}
}
\value{
a data frame with ATT, SE and ranges for lambda and gamma
}
\description{
Function to perform sensitivity analysis of 'killer' confounders
This function takes a data frame, and d and s values as arguments, as defined by Ichino to calculate ATT, SE and ranges for lambda and gamma
}
