% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensbas.R
\name{sensbas}
\alias{sensbas}
\title{Basic function to perform sensitivity analysis in parallel
This function takes bootstrapped data frame and fraction of U=1 by treatment/outcome to calculate ATT and values of lambda and gamma}
\usage{
sensbas(
  trial,
  treatment,
  indvar,
  outvar,
  p00 = 0.5,
  p01 = 0.5,
  p10 = 0.5,
  p11 = 0.5,
  caliperval = 0.1,
  distance = "logit"
)
}
\arguments{
\item{trial}{a rsample bootstrapped dataframe}

\item{treatment}{(\code{character}) Name of treatment variable}

\item{indvar}{(\code{character}) Vector of names of covariate variables}

\item{outvar}{(\code{character}) Name of the outcome variable}

\item{p00}{(\code{numeric}) Value of the probability U=1 when T=0 and Y=0}

\item{p01}{(\code{numeric}) Value of the probability U=1 when for T=0 and Y=1}

\item{p10}{(\code{numeric}) Value of the probability U=1 when for T=1 and Y=0}

\item{p11}{(\code{numeric}) Value of the probability U=1 when for T=1 and Y=1}

\item{caliperval}{(\code{numeric}) Value of the caliper}

\item{distance}{(\code{character}) Type of the distance; "logit" by default}
}
\value{
a list containing ATT value, SE, lambda and gamma values
}
\description{
Basic function to perform sensitivity analysis in parallel
This function takes bootstrapped data frame and fraction of U=1 by treatment/outcome to calculate ATT and values of lambda and gamma
}
